+++ src/server/modules/S3/index.ts
@@ -1,6 +1,7 @@
 import {
+  DeleteObjectCommand,
+  DeleteObjectsCommand,
   GetObjectCommand,
-  ListObjectsCommand,
   PutObjectCommand,
   S3Client,
 } from '@aws-sdk/client-s3';
@@ -45,13 +46,22 @@ export class S3 {
     });
   }
 
-  public async getImages(): Promise<FileType[]> {
-    const command = new ListObjectsCommand({
+  public async deleteFile(key: string) {
+    const command = new DeleteObjectCommand({
       Bucket: this.bucket,
+      Key: key,
     });
 
-    const res = await this.client.send(command);
-    return listFileSchema.parse(res.Contents);
+    return this.client.send(command);
+  }
+
+  public async deleteFiles(keys: string[]) {
+    const command = new DeleteObjectsCommand({
+      Bucket: this.bucket,
+      Delete: { Objects: keys.map((key) => ({ Key: key })) },
+    });
+
+    return this.client.send(command);
   }
 
   public async getFileContent(key: string): Promise<string> {
@@ -67,6 +77,21 @@ export class S3 {
     }
 
     return response.Body.transformToString();
+  }
+
+  public async getFileByteArray(key: string): Promise<Uint8Array> {
+    const command = new GetObjectCommand({
+      Bucket: this.bucket,
+      Key: key,
+    });
+
+    const response = await this.client.send(command);
+
+    if (!response.Body) {
+      throw new Error(`No body in response with ${key}`);
+    }
+
+    return response.Body.transformToByteArray();
   }
 
   public async createPreSignedUrl(key: string): Promise<string> {