+++ src/server/routers/async/index.ts
@@ -1,0 +1,12 @@
+import { publicProcedure, asyncRouter as router } from '@/libs/trpc/async';
+
+import { fileRouter } from './file';
+
+export const asyncRouter = router({
+  file: fileRouter,
+  healthcheck: publicProcedure.query(() => "i'm live!"),
+});
+
+export type AsyncRouter = typeof asyncRouter;
+
+export * from './caller';