+++ src/store/file/slices/chat/action.ts
@@ -1,33 +1,48 @@
-import { t } from 'i18next';
 import { produce } from 'immer';
 import useSWR, { SWRResponse } from 'swr';
 import { StateCreator } from 'zustand/vanilla';
 
-import { notification } from '@/components/AntdStaticMethods';
+import { FILE_UPLOAD_BLACKLIST } from '@/const/file';
 import { fileService } from '@/services/file';
-import { uploadService } from '@/services/upload';
-import { FilePreview } from '@/types/files';
+import { ServerService } from '@/services/file/server';
+import { ragService } from '@/services/rag';
+import { userService } from '@/services/user';
+import { useAgentStore } from '@/store/agent';
+import {
+  UploadFileListDispatch,
+  uploadFileListReducer,
+} from '@/store/file/reducers/uploadFileList';
+import { useUserStore } from '@/store/user';
+import { preferenceSelectors } from '@/store/user/selectors';
+import { FileListItem, FilePreview } from '@/types/files';
+import { UploadFileItem } from '@/types/files/upload';
+import { sleep } from '@/utils/sleep';
 import { setNamespace } from '@/utils/storeDebug';
 
 import { FileStore } from '../../store';
 
-const n = setNamespace('image');
+const n = setNamespace('chat');
+
+const serverFileService = new ServerService();
 
 export interface FileAction {
-  clearImageList: () => void;
+  clearChatUploadFileList: () => void;
+  dispatchChatUploadFileList: (payload: UploadFileListDispatch) => void;
 
-  internal_addFile: (id: string) => void;
-  internal_removeFile: (id: string) => void;
-  internal_toggleLoading: (id: string, loading: boolean) => void;
-  loadFileDetail: (id: string) => Promise<void>;
-  removeAllFiles: () => Promise<void>;
-  removeFile: (id: string) => Promise<void>;
+  removeChatUploadFile: (id: string) => Promise<void>;
+  startAsyncTask: (
+    fileId: string,
+    runner: (id: string) => Promise<string>,
+    onFileItemChange: (fileItem: FileListItem) => void,
+  ) => Promise<void>;
 
-  setImageMapItem: (id: string, item: FilePreview) => void;
+  uploadChatFiles: (files: File[]) => Promise<void>;
 
-  uploadFile: (file: File) => Promise<void>;
+  /**
+   * en: delete it after refactoring the Dalle plugin
+   * @deprecated
+   */
   useFetchFile: (id: string) => SWRResponse<FilePreview>;
-  useFetchFiles: (ids: string[]) => SWRResponse<void>;
 }
 
 export const createFileSlice: StateCreator<
@@ -36,118 +51,158 @@ export const createFileSlice: StateCreator<
   [],
   FileAction
 > = (set, get) => ({
-  clearImageList: () => {
-    set({ inputFilesList: [] }, false, n('clearImageList'));
+  clearChatUploadFileList: () => {
+    set({ chatUploadFileList: [] }, false, n('clearChatUploadFileList'));
   },
-  internal_addFile: (id: string) => {
-    set(({ inputFilesList }) => ({ inputFilesList: [...inputFilesList, id] }), false, n('addFile'));
-  },
-  internal_removeFile: (id: string) => {
-    set(
-      ({ inputFilesList }) => ({ inputFilesList: inputFilesList.filter((i) => i !== id) }),
-      false,
-      n('removeFile'),
-    );
-  },
-  internal_toggleLoading: (id: string, loading) => {
-    if (loading) {
-      set(
-        ({ uploadingIds }) => ({ uploadingIds: [...uploadingIds, id] }),
-        false,
-        'toggleLoading/loading',
-      );
-    } else {
-      set(
-        ({ uploadingIds }) => ({ uploadingIds: uploadingIds.filter((i) => i !== id) }),
-        false,
-        'toggleLoading/stopLoading',
-      );
-    }
-  },
+  dispatchChatUploadFileList: (payload) => {
+    const nextValue = uploadFileListReducer(get().chatUploadFileList, payload);
+    if (nextValue === get().chatUploadFileList) return;
 
-  loadFileDetail: async (id) => {
-    // initFile preview
-    const item = await fileService.getFile(id);
-    get().setImageMapItem(item.id, item);
+    set({ chatUploadFileList: nextValue }, false, `dispatchChatFileList/${payload.type}`);
   },
+  removeChatUploadFile: async (id) => {
+    const { dispatchChatUploadFileList } = get();
 
-  removeAllFiles: async () => {
-    await fileService.removeAllFiles();
-  },
-  removeFile: async (id) => {
-    get().internal_removeFile(id);
-
+    dispatchChatUploadFileList({ id, type: 'removeFile' });
     await fileService.removeFile(id);
   },
 
-  setImageMapItem: (id, item) => {
-    set(
-      produce((draft) => {
-        if (draft.imagesMap[id]) return;
+  startAsyncTask: async (id, runner, onFileItemUpdate) => {
+    await runner(id);
 
-        draft.imagesMap[id] = item;
-      }),
-      false,
-      n('setImageMapItem'),
-    );
+    let isFinished = false;
+
+    while (!isFinished) {
+      // 每间隔 2s 查询一次任务状态
+      await sleep(2000);
+
+      let fileItem: FileListItem | undefined = undefined;
+
+      try {
+        fileItem = await serverFileService.getFileItem(id);
+      } catch (e) {
+        console.error('getFileItem Error:', e);
+        continue;
+      }
+
+      if (!fileItem) return;
+
+      onFileItemUpdate(fileItem);
+
+      if (fileItem.finishEmbedding) {
+        isFinished = true;
+      }
+
+      // if error, also break
+      else if (fileItem.chunkingStatus === 'error' || fileItem.embeddingStatus === 'error') {
+        isFinished = true;
+      }
+    }
   },
 
-  uploadFile: async (file) => {
-    const fileItem = {
-      createdAt: file.lastModified,
-      data: await file.arrayBuffer(),
-      fileType: file.type,
-      name: file.name,
-      saveMode: 'local' as const,
-      size: file.size,
-    };
+  uploadChatFiles: async (rawFiles) => {
+    const { dispatchChatUploadFileList, startAsyncTask } = get();
+    // 0. skip file in blacklist
+    const files = rawFiles.filter((file) => !FILE_UPLOAD_BLACKLIST.includes(file.name));
+    // 1. add files with base64
+    const uploadFiles: UploadFileItem[] = await Promise.all(
+      files.map(async (file) => {
+        let previewUrl: string | undefined = undefined;
+        let base64Url: string | undefined = undefined;
 
-    // at first create a temp id for the file for optimistic rendering
-    const tempId = Date.now().toString();
-    get().internal_addFile(tempId);
+        // only image and video can be previewed, we create a previewUrl and base64Url for them
+        if (file.type.startsWith('image') || file.type.startsWith('video')) {
+          const data = await file.arrayBuffer();
 
-    get().internal_toggleLoading(tempId, true);
+          previewUrl = URL.createObjectURL(new Blob([data!], { type: file.type }));
 
-    // create a local Url for display
-    const url = URL.createObjectURL(new Blob([fileItem.data!], { type: fileItem.fileType }));
-    get().setImageMapItem(tempId, { ...fileItem, id: tempId, url });
+          const base64 = Buffer.from(data!).toString('base64');
+          base64Url = `data:${file.type};base64,${base64}`;
+        }
 
-    try {
-      // after finish upload, mark the `loading=false` to show the uploaded item
-      const result = await uploadService.uploadFile(fileItem);
-      const data = await fileService.createFile(result);
-      get().internal_toggleLoading(tempId, false);
+        return { base64Url, file, id: file.name, previewUrl, status: 'pending' } as UploadFileItem;
+      }),
+    );
 
-      // after finish upload, remove the temp id and add the final one
-      get().internal_removeFile(tempId);
-      get().internal_addFile(data.id);
+    dispatchChatUploadFileList({ files: uploadFiles, type: 'addFiles' });
 
-      // initFile preview
-      await get().loadFileDetail(data.id);
-    } catch (error) {
-      get().internal_removeFile(tempId);
-      get().internal_toggleLoading(tempId, false);
-
-      // show error message
-      notification.error({
-        description: t('upload.desc', { detail: error, ns: 'error' }),
-        message: t('upload.title', { ns: 'error' }),
+    // upload files and process it
+    const pools = files.map(async (file) => {
+      const fileResult = await get().uploadWithProgress({
+        file,
+        onStatusUpdate: dispatchChatUploadFileList,
       });
-    }
+
+      if (!fileResult) return;
+
+      // image don't need to be chunked and embedding
+      if (file.type.startsWith('image')) return;
+
+      // 3. auto chunk and embedding
+      dispatchChatUploadFileList({
+        id: fileResult.id,
+        type: 'updateFile',
+        // make the taks empty to hint the user that the task is starting but not triggered
+        value: { tasks: {} },
+      });
+
+      await startAsyncTask(
+        fileResult.id,
+        async (id) => {
+          const data = await ragService.createParseFileTask(id);
+          if (!data || !data.id) throw new Error('failed to createParseFileTask');
+
+          // run the assignment
+          useAgentStore
+            .getState()
+            .addFilesToAgent([id], false)
+            .then(() => {
+              // trigger the tip if it's the first time
+              if (!preferenceSelectors.shouldTriggerFileInKnowledgeBaseTip(useUserStore.getState()))
+                return;
+
+              userService.updateGuide({ uploadFileInKnowledgeBase: true });
+            });
+
+          return data.id;
+        },
+
+        (fileItem) => {
+          dispatchChatUploadFileList({
+            id: fileResult.id,
+            type: 'updateFile',
+            value: {
+              tasks: {
+                chunkCount: fileItem.chunkCount,
+                chunkingError: fileItem.chunkingError,
+                chunkingStatus: fileItem.chunkingStatus,
+                embeddingError: fileItem.embeddingError,
+                embeddingStatus: fileItem.embeddingStatus,
+                finishEmbedding: fileItem.finishEmbedding,
+              },
+            },
+          });
+        },
+      );
+    });
+
+    await Promise.all(pools);
   },
 
   useFetchFile: (id) =>
     useSWR(id, async (id) => {
       const item = await fileService.getFile(id);
 
-      get().setImageMapItem(id, item);
+      set(
+        produce((draft) => {
+          if (draft.imagesMap[id]) return;
+
+          draft.imagesMap[id] = item;
+        }),
+        false,
+        n('useFetchFile'),
+      );
 
       return item;
-    }),
-  useFetchFiles: (ids) =>
-    useSWR(ids, async (ids) => {
-      const pools = ids.map(async (id) => get().loadFileDetail(id));
-
-      await Promise.all(pools);
     }),
 });