+++ src/store/file/slices/chat/initialState.ts
@@ -1,13 +1,18 @@
 import { FilePreview } from '@/types/files';
+import { UploadFileItem } from '@/types/files/upload';
 
 export interface ImageFileState {
+  chatUploadFileList: UploadFileItem[];
+  /**
+   * it should be removed after dalle plugin refactor
+   * @deprecated
+   */
   imagesMap: Record<string, FilePreview>;
-  inputFilesList: string[];
   uploadingIds: string[];
 }
 
 export const initialImageFileState: ImageFileState = {
+  chatUploadFileList: [],
   imagesMap: {},
-  inputFilesList: [],
   uploadingIds: [],
 };