+++ src/store/knowledgeBase/selectors.ts
@@ -1,0 +1,1 @@
+export * from './slices/crud/selectors';